Revert "Polished Boss and Side Character"
