+++ goblin.py
@@ -1,8 +1,28 @@
-from enemy import enemy
 import random
 
-class Goblin(enemy):
+class Goblin:
+    """
+    This is our goblin blueprint 
+    
+    Attributes:
+        name: Awe, it has a name? How cute!
+        health: The current health value 
+        attack_power: How much health will be drained from opponent if hit
+    """
+    def __init__(self, name):
+        self.name = name
+        self.health = random.randint(584,793)
+        self.attack_power = random.randint(145, 175)
 
-    def __init__(self, name, color):
-        super().__init__(name)
-        self.color = color
+    def attack(self):
+        return random.randint(1, self.attack_power)
+
+    def take_damage(self, damage):
+        self.health -= damage
+        if self.health <= 0:
+            self.health = 0
+            return self.health
+        print(f"{self.name} takes {damage} damage. Health is now {self.health}.")
+
+    def is_alive(self):
+        return self.health > 0
